backend/requirements.txt

fastapi>=0.111.0
uvicorn[standard]>=0.30.1
python-multipart>=0.0.9
pydantic>=2.9.0
pydantic-settings>=2.5.0

# LangChain
langchain>=0.2.5
langchain-google-genai>=1.0.6
langchain-community>=0.2.5
langchain-chroma>=1.0.0

# Vector DB
chromadb>=0.5.3

# Document parsing
pypdfium2>=4.30.0
python-docx>=1.1.2

# Google Gemini
google-generativeai>=0.7.2

# Utilities
python-dotenv>=1.0.1
aiofiles>=23.2.1